+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### [Unreleased][unreleased]
 
 #### Added
-- json output for init command
+- json output for commands
+- migration datetime to Status command
 
 ### [0.7.0] - 2017-04-06
 #### Changed